+++ Js2IL.Tests/Node/Buffer/ExecutionTests.cs
@@ -1,0 +1,13 @@
+using System.Threading.Tasks;
+
+namespace Js2IL.Tests.Node.Buffer
+{
+    public class ExecutionTests : ExecutionTestsBase
+    {
+        public ExecutionTests() : base("Node/Buffer") { }
+
+        [Fact]
+        public Task Buffer_From_And_IsBuffer()
+            => ExecutionTest(nameof(Buffer_From_And_IsBuffer));
+    }
+}